+++ pkg/cmd/kafka/kafkacmdutl/kafka_util.go
